Extract opening element and closing element visitors into helper functions
